fix/Переместил партнеров
