--- src/components/main-page/partners/index.ts
@@ -1,1 +1,0 @@
-export * from './mainPartners';
