--- src/components/main-page/partners/mainPartners.module.css
@@ -1,43 +1,0 @@
-.partners {
-  width: calc(100% - scale(270px));
-  box-sizing: border-box;
-  padding: scale(106px 60px 160px);
-  background-color: var(--beige);
-  grid-area: partners;
-
-  @media (max-width: $tablet-portrait) {
-    width: 100%;
-    padding: scale(56px 24px) 0;
-  }
-}
-
-.link {
-  text-decoration: none;
-}
-
-.wrapper {
-  position: relative;
-  width: 100%;
-  height: 100%;
-  max-height: scale(83px);
-  transition: all .3s;
-
-  &:hover {
-    opacity: .7;
-  }
-
-  @media (max-width: $tablet-portrait) {
-    max-width: scale(160px);
-    max-height: scale(50px);
-    margin: auto;
-  }
-}
-
-.text {
-  @mixin text;
-  @mixin textPartners;
-
-  text-align: center;
-  margin: 0;
-  margin: scale(20px 0 15px);
-}
