+++ src/components/partners-layout/index.ts
@@ -1,0 +1,1 @@
+export * from './partners-layout';
