+++ src/components/partners/partners.module.css
@@ -1,39 +1,43 @@
 .partners {
-  margin-bottom: scale(70px);
-
-  &:last-of-type {
-    margin-bottom: 0;
-  }
-}
-
-.title {
-  @mixin headline;
-  @mixin headline6;
-
-  max-width: scale(334px);
-  margin-bottom: scale(16px);
+  width: calc(100% - scale(270px));
+  box-sizing: border-box;
+  padding: scale(106px 60px 160px);
+  background-color: var(--beige);
+  grid-area: partners;
 
   @media (max-width: $tablet-portrait) {
-    margin-bottom: scale(24px);
+    width: 100%;
+    padding: scale(56px 24px) 0;
   }
 }
 
-.list {
-  display: grid;
-  padding: 0;
+.link {
+  text-decoration: none;
+}
+
+.wrapper {
+  position: relative;
+  width: 100%;
+  height: 100%;
+  max-height: scale(83px);
+  transition: all .3s;
+
+  &:hover {
+    opacity: .7;
+  }
+
+  @media (max-width: $tablet-portrait) {
+    max-width: scale(160px);
+    max-height: scale(50px);
+    margin: auto;
+  }
+}
+
+.text {
+  @mixin text;
+  @mixin textPartners;
+
+  text-align: center;
   margin: 0;
-  margin-left: scale(60px);
-  column-gap: scale(30px);
-  grid-auto-rows: scale(138px);
-  grid-template-columns: repeat(auto-fill, scale(210px));
-  list-style-type: none;
-  row-gap: scale(16px);
-
-  @media (max-width: $tablet-portrait) {
-    margin-left: 0;
-    column-gap: 0;
-    grid-auto-rows: scale(118px);
-    grid-template-columns: repeat(2, minmax(scale(170px), scale(182px)));
-    row-gap: 0;
-  }
+  margin: scale(20px 0 15px);
 }
